+++ app/src/main/java/com/apps/kunalfarmah/echo/util/Constants.java
@@ -5,10 +5,8 @@ public class Constants {
 
     public interface ACTION {
         String MAIN_ACTION = "com.apps.kunalfarmah.action.main";
-        String INIT_ACTION = "com.apps.kunalfarmah.action.init";
         String PREV_ACTION = "com.apps.kunalfarmah.action.prev";
         String PLAY_ACTION = "com.apps.kunalfarmah.action.play";
-        String PAUSE_ACTION = "com.apps.kunalfarmah.action.pause";
         String NEXT_ACTION = "com.apps.kunalfarmah.action.next";
         String SHUFFLE_ACTION = "com.apps.kunalfarmah.action.shuffle";
         String STARTFOREGROUND_ACTION = "com.apps.kunalfarmah.action.startforeground";
@@ -16,9 +14,7 @@ public class Constants {
         String CHANGE_TO_PAUSE = "com.apps.kunalfarmah.action.changetopause";
         String CHANGE_TO_PLAY = "com.apps.kunalfarmah.action.changetoplay";
         String NEXT_UPDATE = "com.apps.kunalfarmah.action.nextupdate";
-        String NEXT_UPDATE_SHUFFLE = "com.apps.kunalfarmah.action.nextupdateshuffle";
         String PREV_UPDATE = "com.apps.kunalfarmah.action.prevupdate";
-        String PREV_UPDATE_SHUFFLE = "com.apps.kunalfarmah.action.prevupdateshuffle";
         String CLOSE = "com.apps.kunalfarmah.action.close";
     }
 
@@ -32,7 +28,6 @@ public class Constants {
 
     public interface NOTIFICATION_ID {
         int FOREGROUND_SERVICE = 101;
-        int PLAYER_NOTIFICATION_MANAGER = 102;
     }
 
 }
