--- app/src/main/res/layout/fragment_about.xml
@@ -1,178 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:background="@color/colorPrimary"
-    android:padding="@dimen/_5sdp"
-    android:clickable="true"
-    tools:context="com.apps.kunalfarmah.echo.fragment.AboutUSFragment">
-
-    <TextView
-        android:id="@+id/dev"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="center_horizontal"
-        android:padding="@dimen/_5sdp"
-        android:text="YOUR DEVELOPER"
-        android:textColor="#ffffff"
-        android:layout_marginBottom="@dimen/_35sdp"
-        android:textAlignment="center"
-        android:textSize="@dimen/_25sdp" />
-
-    <TextView
-        android:id="@+id/info1"
-        android:visibility="gone"
-        android:layout_below="@+id/dev"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/_5sdp"
-        android:text="This App has been designed by a first time Android Developer KUNAL FARMAH."
-        android:textSize="@dimen/_15sdp" />
-
-    <TextView
-        android:layout_marginTop="@dimen/_5sdp"
-        android:id="@+id/we"
-        android:visibility="gone"
-        android:layout_below="@+id/info1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="@dimen/_15sdp"
-        android:text="Currently I am persuing B.Tech in Information Technology from Delhi Technological University"/>
-    <TextView
-        android:id="@+id/info2"
-        android:visibility="gone"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_below="@+id/we"
-        android:layout_marginBottom="7dp"
-        android:layout_marginTop="@dimen/_5sdp"
-        android:text="I have tested this app to the best of my knowlege.
-If anyone finds a bug or a crash at some point, feel free to reach me at the undersigned"
-        android:textSize="@dimen/_15sdp" />
-
-
-
-    <TextView
-        android:id="@+id/Name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/dev"
-        android:layout_marginTop="@dimen/_20sdp"
-        android:layout_alignParentEnd="true"
-        android:layout_marginEnd="48dp"
-        android:padding="@dimen/_5sdp"
-        android:textColor="#ffffff"
-        android:text="KUNAL FARMAH"
-        android:textSize="@dimen/_15sdp" />
-
-    <TextView
-        android:id="@+id/ADD"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/Name"
-        android:layout_alignStart="@+id/Name"
-        android:padding="@dimen/_5sdp"
-        android:textColor="#ffffff"
-        android:text="NEW DELHI, INDIA"
-        android:textSize="@dimen/_15sdp" />
-
-
-    <ImageView
-        android:id="@+id/me"
-        android:layout_width="@dimen/_134sdp"
-        android:layout_height="@dimen/_104sdp"
-        android:layout_alignParentStart="true"
-        android:layout_below="@id/info2"
-        android:layout_marginTop="@dimen/_5sdp"
-        android:src="@drawable/me" />
-
-
-    <LinearLayout
-
-        android:id="@+id/emailinfo"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/me"
-        android:layout_marginTop="@dimen/_5sdp"
-        android:layout_centerHorizontal="true"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:layout_marginLeft="@dimen/_10sdp"
-            android:id="@+id/mailme"
-            android:layout_gravity="end"
-            android:layout_width="@dimen/_30sdp"
-            android:layout_height="@dimen/_30sdp"
-            android:layout_alignBottom="@+id/email"
-            android:layout_toStartOf="@+id/dev"
-            android:src="@drawable/gmail1" />
-
-        <TextView
-            android:id="@+id/email"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:clickable="true"
-            android:layout_above="@+id/reachme"
-            android:layout_alignEnd="@+id/extra1"
-            android:layout_marginLeft="@dimen/_15sdp"
-            android:gravity="center"
-            android:textColor="#ffffff"
-            android:text="kunalfarmah98@gmail.com"
-            android:textSize="@dimen/_15sdp" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/fbinfo"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/emailinfo"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="@dimen/_5sdp"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:id="@+id/reachme"
-            android:layout_marginLeft="@dimen/_10sdp"
-            android:layout_width="@dimen/_30sdp"
-            android:layout_height="@dimen/_30sdp"
-            android:layout_alignParentBottom="true"
-            android:layout_alignStart="@+id/mailme"
-            android:src="@drawable/fb1" />
-
-        <TextView
-            android:id="@+id/fb"
-            android:layout_width="wrap_content"
-            android:clickable="true"
-            android:layout_height="match_parent"
-            android:layout_above="@+id/reachme"
-            android:layout_alignEnd="@+id/extra1"
-            android:layout_marginLeft="@dimen/_15sdp"
-            android:gravity="center"
-            android:textColor="#ffffff"
-            android:text="www.facebook.com/kunal.farmah"
-            android:textSize="@dimen/_15sdp" />
-
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/smile"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:padding="@dimen/_5sdp"
-        android:layout_below="@id/fbinfo"
-        android:layout_marginTop="@dimen/_10sdp"
-        android:text="Cheers to ALL :)"
-        android:textAlignment="center"
-        android:textStyle="italic"
-        android:textColor="#ffffff"
-        android:layout_centerHorizontal="true"
-        android:textSize="@dimen/_20sdp" />
-
-
-</RelativeLayout>
